+++ src/components/utils/header.jsx
@@ -21,7 +21,7 @@ export const Header = ({choice}) => {
     };
 
     return (
-        <header className="inset-x-0 top-0 bg-neutral-800 h-14 w-full border-y border-zinc-600 z-50">
+        <header className="inset-x-0 top-0 bg-neutral-800 h-14 w-full border-y border-zinc-600">
             <section className="absolute top-2 left-0">
                 <img onClick={handleClick} draggable='false' src={logo} className="-mt-2.5 cursor-pointer ml-2.5 scale-75 rounded-full" alt="logo" />
             </section>